# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SL Real Estate & Construction Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /properties:
    get:
      operationId: getProperties
      tags: [properties]
      summary: Get all approved properties
      parameters:
        - name: district
          in: query
          schema:
            type: string
        - name: listingType
          in: query
          schema:
            type: string
            enum: [sale, rent]
        - name: propertyType
          in: query
          schema:
            type: string
            enum: [house, land, apartment, commercial]
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: minArea
          in: query
          schema:
            type: number
        - name: maxArea
          in: query
          schema:
            type: number
        - name: bedrooms
          in: query
          schema:
            type: number
        - name: amenities
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: number
        - name: offset
          in: query
          schema:
            type: number
      responses:
        "200":
          description: List of properties
          content:
            application/json:
              schema:
                type: object
                properties:
                  properties:
                    type: array
                    items:
                      $ref: "#/components/schemas/Property"
                  total:
                    type: number
                required:
                  - properties
                  - total

    post:
      operationId: createProperty
      tags: [properties]
      summary: Submit a new property listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePropertyInput"
      responses:
        "201":
          description: Property created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /properties/{id}:
    get:
      operationId: getProperty
      tags: [properties]
      summary: Get a property by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Property details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      operationId: updateProperty
      tags: [properties]
      summary: Update a property (owner or admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePropertyInput"
      responses:
        "200":
          description: Updated property
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"

    delete:
      operationId: deleteProperty
      tags: [properties]
      summary: Delete a property
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
  /admin/properties:
    get:
      operationId: adminGetProperties
      tags: [admin]
      summary: Admin - get all properties including pending
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected]
        - name: limit
          in: query
          schema:
            type: number
        - name: offset
          in: query
          schema:
            type: number
      responses:
        "200":
          description: All properties
          content:
            application/json:
              schema:
                type: object
                properties:
                  properties:
                    type: array
                    items:
                      $ref: "#/components/schemas/Property"
                  total:
                    type: number
                required:
                  - properties
                  - total
...
  /services:
    get:
      operationId: getServices
      tags: [services]
      summary: Get all construction services
      parameters:
        - name: district
          in: query
          schema:
            type: string
        - name: serviceType
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of services
          content:
            application/json:
              schema:
                type: object
                properties:
                  services:
                    type: array
                    items:
                      $ref: "#/components/schemas/ConstructionService"
                required:
                  - services

    post:
      operationId: createService
      tags: [services]
      summary: Admin - create a construction service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceInput"
      responses:
        "201":
          description: Service created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConstructionService"
  /services/{id}:
    put:
      operationId: updateService
      tags: [services]
      summary: Admin - update a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceInput"
      responses:
        "200":
          description: Updated service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConstructionService"

    delete:
      operationId: deleteService
      tags: [services]
      summary: Admin - delete a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    Property:
      type: object
      properties:
        id:
          type: number
        title:
          type: string
        description:
          type: string
        listingType:
          type: string
          enum: [sale, rent]
        propertyType:
          type: string
          enum: [house, land, apartment, commercial]
        district:
          type: string
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        priceNpr:
          type: number
        areaDhur:
          type: number
          nullable: true
        bedrooms:
          type: number
          nullable: true
        bathrooms:
          type: number
          nullable: true
        buildYear:
          type: number
          nullable: true
        amenities:
          type: array
          items:
            type: string
        photos:
          type: array
          items:
            type: string
        videoUrl:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, approved, rejected]
        featured:
          type: boolean
        ownerName:
          type: string
        ownerPhone:
          type: string
        ownerWhatsapp:
          type: string
          nullable: true
        rejectionReason:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - title
        - description
        - listingType
        - propertyType
        - district
        - priceNpr
        - amenities
        - photos
        - status
        - featured
        - ownerName
        - ownerPhone
        - createdAt
        - updatedAt
    CreatePropertyInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        listingType:
          type: string
          enum: [sale, rent]
        propertyType:
          type: string
          enum: [house, land, apartment, commercial]
        district:
          type: string
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        priceNpr:
          type: number
        areaDhur:
          type: number
          nullable: true
        bedrooms:
          type: number
          nullable: true
        bathrooms:
          type: number
          nullable: true
        buildYear:
          type: number
          nullable: true
        amenities:
          type: array
          items:
            type: string
        photos:
          type: array
          items:
            type: string
        videoUrl:
          type: string
          nullable: true
        ownerName:
          type: string
        ownerPhone:
          type: string
        ownerWhatsapp:
          type: string
          nullable: true
      required:
        - title
        - description
        - listingType
        - propertyType
        - district
        - priceNpr
        - ownerName
        - ownerPhone
    UpdatePropertyInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        listingType:
          type: string
          enum: [sale, rent]
        propertyType:
          type: string
          enum: [house, land, apartment, commercial]
        district:
          type: string
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        priceNpr:
          type: number
        areaDhur:
          type: number
          nullable: true
        bedrooms:
          type: number
          nullable: true
        bathrooms:
          type: number
          nullable: true
        buildYear:
          type: number
          nullable: true
        amenities:
          type: array
          items:
            type: string
        photos:
          type: array
          items:
            type: string
        videoUrl:
          type: string
          nullable: true
        ownerName:
          type: string
        ownerPhone:
          type: string
        ownerWhatsapp:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, approved, rejected]
        featured:
          type: boolean
    ConstructionService:
      type: object
      properties:
        id:
          type: number
        businessName:
          type: string
        serviceType:
          type: string
        phone:
          type: string
        whatsapp:
          type: string
          nullable: true
        district:
          type: string
        description:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - businessName
        - serviceType
        - phone
        - district
        - createdAt
    CreateServiceInput:
      type: object
      properties:
        businessName:
          type: string
        serviceType:
          type: string
        phone:
          type: string
        whatsapp:
          type: string
          nullable: true
        district:
          type: string
        description:
          type: string
          nullable: true
      required:
        - businessName
        - serviceType
        - phone
        - district
...