# lib/api-zod/src/generated/types/constructionService.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SL Real Estate & Construction Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface ConstructionService {
  id: number;
  businessName: string;
  serviceType: string;
  phone: string;
  whatsapp?: string | null;
  district: string;
  description?: string | null;
  createdAt: string;
}